requirements.txt
absl-py==1.4.0
affine==2.4.0
albumentations==1.3.1
alembic==1.11.3
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
augmentations==0.0.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.7.0
bleach==6.0.0
blinker==1.6.2
build==0.10.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.5
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cmake==3.27.2
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cookiecutter==2.2.3
cycler==0.11.0
Cython==3.0.0
databricks-cli==0.17.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dlib==19.24.2
docker==6.1.3
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
elasticdeform==0.5.0
entrypoints==0.4
evalutils==0.4.2
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.3.3
flatbuffers==23.5.26

fonttools==4.41.0
fqdn==1.5.1
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gast==0.4.0
gdcm==1.1
geojson==3.0.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.57.0
gunicorn==21.2.0
h5py==3.9.0
huggingface-hub==0.20.1
humanize==4.7.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
imutils==0.5.4
iopath==0.1.9
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter-console==6.6.3
jupyter_core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab-widgets==3.0.8
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
lit==16.0.6
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.1
mlflow==2.6.0
monai==1.3.0
mpmath==1.3.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nibabel==5.1.0
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.3.1
pip-tools==7.0.0
pipreqs==0.4.13
platformdirs==3.10.0
plotly==5.18.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydicom==2.4.3
Pygments==2.16.1
PyJWT==2.8.0
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.4
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
querystring-parser==1.2.4
rasterio==1.3.8
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.0
Send2Trash==1.8.2
sentry-sdk==1.39.1
setproctitle==1.3.3
setuptools==58.1.0
shapely==2.0.1
shellingham==1.5.0.post1
SimpleITK==2.2.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.0.dev20231112+cu118
torch-tb-profiler==0.4.3
torchio==0.19.1
torchsummary==1.5.1
torchvision==0.17.0.dev20231112+cu118
torchxrayvision==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trimesh==3.22.4
triton==2.1.0
turfpy==0.0.7
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.200
uri-template==1.3.0
urllib3==1.26.16
wandb==0.16.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.7
wheel==0.41.1
widgetsnbextension==4.0.8
wrapt==1.15.0
yacs==0.1.8
yarg==0.1.9
zipp==3.16.2
